+++ pragma-node/src/handlers/onchain/mod.rs
@@ -1,5 +1,5 @@
 pub mod get_checkpoints;
 pub mod get_entry;
 pub mod get_history;
-pub mod get_ohlc;
 pub mod get_publishers;
+pub mod subscribe_to_ohlc;
