+++ pragma-node/src/servers/app/mod.rs
@@ -40,7 +40,7 @@ pub async fn run_app_server(config: &Config, state: AppState) {
             handlers::onchain::get_history::get_onchain_history,
             handlers::onchain::get_checkpoints::get_onchain_checkpoints,
             handlers::onchain::get_publishers::get_onchain_publishers,
-            handlers::onchain::get_ohlc::subscribe_to_onchain_ohlc,
+            handlers::onchain::subscribe_to_ohlc::subscribe_to_onchain_ohlc,
             handlers::merkle_feeds::get_option::get_merkle_feeds_option,
             handlers::merkle_feeds::get_merkle_proof::get_merkle_feeds_proof,
         ),
@@ -64,7 +64,7 @@ pub async fn run_app_server(config: &Config, state: AppState) {
                 handlers::onchain::get_checkpoints::GetOnchainCheckpointsResponse,
                 handlers::onchain::get_publishers::GetOnchainPublishersParams,
                 handlers::onchain::get_publishers::GetOnchainPublishersResponse,
-                handlers::onchain::get_ohlc::GetOnchainOHLCResponse,
+                handlers::onchain::subscribe_to_ohlc::GetOnchainOHLCResponse,
                 handlers::onchain::get_history::GetOnchainHistoryParams,
                 handlers::onchain::get_history::GetOnchainHistoryResponse,
                 handlers::merkle_feeds::get_option::GetOptionQuery,
