+++ pragma-node/src/servers/app/routes.rs
@@ -10,8 +10,8 @@ use crate::handlers::merkle_feeds::{
 };
 use crate::handlers::onchain::{
     get_checkpoints::get_onchain_checkpoints, get_entry::get_onchain_entry,
-    get_history::get_onchain_history, get_ohlc::subscribe_to_onchain_ohlc,
-    get_publishers::get_onchain_publishers,
+    get_history::get_onchain_history, get_publishers::get_onchain_publishers,
+    subscribe_to_ohlc::subscribe_to_onchain_ohlc,
 };
 use crate::handlers::{
     create_entries, create_future_entries, get_entry, get_expiries, get_ohlc, get_volatility,
